Handling Null Exceptions
Made CustomerType constructor instantiate an array of DvdType that is all initialized to null (less programming in the main). 
Edited Rent and Return to handle null objects and deal with it accordingly. You can't return null. If you find a null spot in the array, put the DVD in that array.
Fixed printDVD method to return titles and not a messy DVD toString.
Demonstrated changes in main. Look at customers using rent method